microservices/auth-service/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pydantic
requests
sqlalchemy
psycopg2-binary
PyJWT
bcrypt

